Making the docker deb packages linux only
