+++ CMakeLists.txt
@@ -257,7 +257,7 @@ else()
     )
     if (CMAKE_CXX_COMPILER MATCHES "clang" AND NOT FREEBSD)
       # Only enable LTO builds when using clang on Unix, for now
-      add_compile_options(-flto=thin)
+      # add_compile_options(-flto=thin)
     endif()
   endif()
 endif()
@@ -403,8 +403,8 @@ if(WINDOWS)
 elseif(NOT FREEBSD)
   if (CMAKE_CXX_COMPILER MATCHES "clang")
     # Clang on Unix uses LTO; we also need to pass -flto when linking
-    set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -flto=thin")
-    set(CMAKE_SHARED_LINKER_FLAGS "${CMAKE_SHARED_LINKER_FLAGS} -flto=thin")
+    # set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -flto=thin")
+    # set(CMAKE_SHARED_LINKER_FLAGS "${CMAKE_SHARED_LINKER_FLAGS} -flto=thin")
   endif()
 endif()
