+++ osquery/tables/applications/posix/docker_api.h
@@ -1,0 +1,12 @@
+#include <boost/property_tree/ptree.hpp>
+
+#include <osquery/tables.h>
+
+namespace pt = boost::property_tree;
+
+namespace osquery {
+namespace tables {
+bool checkConstraintValue(const std::string& str);
+Status dockerApi(const std::string& uri, pt::ptree& tree);
+}
+}
