+++ osquery/tables/system/linux/deb.cpp
@@ -1,13 +1,13 @@
 #include <osquery/tables/system/linux/deb.h>
 
-
 namespace osquery {
+namespace tables {
 
-int pkg_sorter(const void *a, const void *b) {
-  const struct pkginfo *pa = *(const struct pkginfo **)a;
-  const struct pkginfo *pb = *(const struct pkginfo **)b;
-  const char *arch_a = pa->installed.arch->name;
-  const char *arch_b = pb->installed.arch->name;
+int pkg_sorter(const void* a, const void* b) {
+  const struct pkginfo* pa = *(const struct pkginfo**)a;
+  const struct pkginfo* pb = *(const struct pkginfo**)b;
+  const char* arch_a = pa->installed.arch->name;
+  const char* arch_b = pb->installed.arch->name;
 
   int res = strcmp(pa->set->name, pb->set->name);
   if (res != 0) {
@@ -21,11 +21,11 @@ int pkg_sorter(const void *a, const void *b) {
   return strcmp(arch_a, arch_b);
 }
 
-void w_revision(struct varbuf *vb,
-                const struct pkginfo *pkg,
-                const struct pkgbin *pkgbin,
+void w_revision(struct varbuf* vb,
+                const struct pkginfo* pkg,
+                const struct pkgbin* pkgbin,
                 enum fwriteflags flags,
-                const struct fieldinfo *fip) {
+                const struct fieldinfo* fip) {
   if (flags & fw_printheader) {
     varbuf_add_str(vb, "Revision: ");
   }
@@ -35,7 +35,7 @@ void w_revision(struct varbuf *vb,
   }
 }
 
-void dpkg_setup(struct pkg_array *packages) {
+void dpkg_setup(struct pkg_array* packages) {
   dpkg_set_progname("osquery");
   push_error_context();
 
@@ -47,12 +47,13 @@ void dpkg_setup(struct pkg_array *packages) {
   pkg_array_sort(packages, pkg_sorter);
 }
 
-void dpkg_teardown(struct pkg_array *packages) {
+void dpkg_teardown(struct pkg_array* packages) {
   pkg_array_destroy(packages);
 
   pkg_db_reset();
   modstatdb_done();
 
   pop_error_context(ehflag_normaltidy);
+}
 }
 }
